Add first CAM implementation
This jupyter notebook includes an implementation of a Resnet18 convolutional
neural network. This network is trained with CIFAR-10 dataset and gets an
accuracy of 93% (approx).

The information of this CNN is used to compute a " class activation map (CAM)".
The CAM help us to identify the discriminative image regions used by the CNN to
identify a certain category.

Signed-off-by: mario carrillo <carrillo@rhrk.uni-kl.de>